This fix will change the tab to the relay view, wait half a second, and then go to the profile view
