+++ Nos/Views/HomeFeedView.swift
@@ -74,8 +74,8 @@ struct HomeFeedView: View {
             .navigationDestination(for: Author.self) { author in
                 ProfileView(author: author)
             }
-            .navigationDestination(for: AppView.Destination.self) { destination in
-                if destination == AppView.Destination.settings {
+            .navigationDestination(for: Destination.self) { destination in
+                if destination == Destination.settings {
                     SettingsView()
                 }
             }
