+++ Nos/Views/RelayView.swift
@@ -75,8 +75,8 @@ struct RelayView: View {
             .navigationDestination(for: Author.self) { author in
                 ProfileView(author: author)
             }
-            .navigationDestination(for: AppView.Destination.self) { destination in
-                if destination == AppView.Destination.settings {
+            .navigationDestination(for: Destination.self) { destination in
+                if destination == Destination.settings {
                     SettingsView()
                 }
             }
